[feat] round table에 승점 로직 추가
